moves images, tweaks css, moves cookies loc (erring)
