+++ src/components/NavGlobal.tsx
@@ -1,52 +1,47 @@
-import React from 'react';
 import Image from 'next/image';
-// import cloudIcon from '@/../public/img/cloud-gov-icon.svg';
-import cloudGovLogo from '@/../public/img/cloud-gov-logo-white.svg';
-import dashboardIcon from '@/../public/img/dashboard-icon.svg';
-import cloudPagesIcon from '@/../public/img/cloud-pages-icon.svg';
 import Link from 'next/link';
+import { cookies } from 'next/headers';
+
+import cloudGovText from '@/../public/img/logos/cloud-gov-logo-text.svg';
+import cloudGovIcon from '@/../public/img/logos/cloud-gov-icon.svg';
+import dashboardIcon from '@/../public/img/logos/dashboard-icon.svg';
+import cloudPagesIcon from '@/../public/img/logos/cloud-pages-icon.svg';
+
+import { LoginButton } from '@/components/auth/LoginButton';
+import { LogoutButton } from '@/components/auth/LogoutButton';
 
 export function NavGlobal() {
+  const cookieStore = cookies();
+  const authSession = cookieStore.get('authsession');
+
   return (
-    // add height-8 if you want it to be up to the spec
     <div id="nav-global">
       <div>
         <div className="nav-global-inner">
-          <Link href="/" className="margin-right-7">
+          <Link href="/" className="cloud-logo">
             <Image
-              id="cloud-gov-logo"
+              id="cloud-gov-icon"
               unoptimized
-              src={cloudGovLogo}
+              src={cloudGovIcon}
+              alt="cloud.gov"
+            />
+            <Image
+              id="cloud-gov-logo-text"
+              unoptimized
+              src={cloudGovText}
               alt="cloud.gov"
             />
           </Link>
-          <span className="products"></span>
-          <Link href="/">
-            <span className="nav-product active">
-              <Image
-                unoptimized
-                src={dashboardIcon}
-                alt="dashboard"
-                height={24}
-                width={24}
-              />
-              <span>Dashboard</span>
-            </span>
+          <Link href="/" className="nav-product active">
+            <Image unoptimized src={dashboardIcon} alt="dashboard" />
+            <span>Dashboard</span>
           </Link>
-          <Link href="/">
-            <span className="nav-product">
-              <Image
-                unoptimized
-                src={cloudPagesIcon}
-                alt="Cloud Pages icon"
-                height={24}
-                width={24}
-              />
-              <span>Pages</span>
-            </span>
+          <Link href="/" className="nav-product">
+            <Image unoptimized src={cloudPagesIcon} alt="Cloud Pages icon" />
+            <span>Pages</span>
           </Link>
           <span className="margin-left-auto margin-bottom-05 border-bottom-05 border-transparent">
-            Log out
+            {authSession ? <LogoutButton /> : <LoginButton />}
           </span>
         </div>
       </div>
